Separate the script into different parts
See: #12
